Invoke child_process directly for chromedriver
`chromedriver.start` doesn't have the ability to run the process as detached, which means it automatically kills chromedriver if the parent process receives a SIGINT. This prevents the process from being able to exit gracefully if using the internal chromedriver process.

Instead call `cp.spawn` directly with the chromedriver binary path, and set `detached` option to true.